adds active league section
