+++ src/screens/JoinLeagueScreen.tsx
@@ -14,6 +14,9 @@ import { useCountdown } from "../hooks/useCountdown";
 import { useUserFantasyTeam } from "../components/league/useFantasyLeague";
 import { useRouter } from "../hooks/useRoter";
 import { twMerge } from "tailwind-merge";
+import JoinLeagueDeadlineCountdown from "../components/leagues/JoinLeagueDeadlineContdown";
+import { ActiveLeaguesSection } from "../components/dashboard";
+import JoinLeagueActiveLeaguesSection from "../components/leagues/join_league_screen/JoinLeagueActiveLeaguesSection";
 
 export function JoinLeagueScreen() {
   const navigate = useNavigate();
@@ -35,9 +38,6 @@ export function JoinLeagueScreen() {
   const leaguesOnTheClock = activeLeagues.filter((l) => {
     return (isLeagueOnTheClock(l, twoDays));
   });
-
-
-  console.log(leaguesOnTheClock);
 
   if (leaguesOnTheClock.length > 0) {
     leagueOnTheClock = leaguesOnTheClock[0];
@@ -63,11 +63,13 @@ export function JoinLeagueScreen() {
         const leagues = await leagueService.getAllLeagues();
 
         // Filter leagues based on is_open status
-        const available = leagues.filter(
-          (league) => league.is_open && !league.has_ended
-        );
+        // const available = leagues.filter(
+        //   (league) => league.is_open && !league.has_ended
+        // );
 
-        setAvailableLeagues(available);
+
+
+        setAvailableLeagues(leagues);
         setError(null);
       } catch (err) {
         console.error("Failed to fetch leagues:", err);
@@ -148,7 +150,7 @@ export function JoinLeagueScreen() {
     });
   };
 
-  const leagues = activeLeaguesFilter(availableLeagues)
+  const leagues = availableLeagues;
 
   const otherLeagues = availableLeagues
     .filter(l => l.join_deadline === null);
@@ -161,7 +163,7 @@ export function JoinLeagueScreen() {
         </h1>
       </div>
 
-      {leagueOnTheClock && <JoinDeadlineCountdown league={leagueOnTheClock} onViewLeague={handleLeagueClick} />}
+      {leagueOnTheClock && <JoinLeagueDeadlineCountdown league={leagueOnTheClock} onViewLeague={handleLeagueClick} />}
 
       {isLoading ? (
         <div className="flex justify-center items-center py-12">
@@ -191,31 +193,12 @@ export function JoinLeagueScreen() {
         </div>
       ) : (
         <>
-          {!isLoading && <div className="bg-white dark:bg-gray-800/40 rounded-xl shadow-sm my-6 p-4 sm:p-6">
-            <h2 className="text-xl font-semibold flex items-center gap-2 mb-6 dark:text-gray-100">
-              <Trophy size={24} className="text-primary-500" />
-              Active Leagues
-            </h2>
-
-            <motion.div
-              variants={containerVariants}
-              initial="hidden"
-              animate="visible"
-              className="space-y-4"
-            >
-              {leagues.map((league, index) => (
-                <LeagueCard
-                  key={league.id}
-                  league={league}
-                  onLeagueClick={handleLeagueClick}
-                  teamCount={teamCounts[league.id]}
-                  isLoading={isLoadingCounts}
-                  custom={index}
-                  isJoined={userTeams[league.id]}
-                />
-              ))}
-            </motion.div>
-          </div>}
+          {!isLoading && 
+            <JoinLeagueActiveLeaguesSection
+              leagues={leagues}
+              userTeams={userTeams}
+            />
+          }
         </>
       )}
 
@@ -237,8 +220,6 @@ export function JoinLeagueScreen() {
               key={league.id}
               league={league}
               onLeagueClick={handleLeagueClick}
-              teamCount={teamCounts[league.id]}
-              isLoading={isLoadingCounts}
               custom={index}
               isJoined={userTeams[league.id]}
             />
@@ -250,82 +231,3 @@ export function JoinLeagueScreen() {
   );
 }
 
-type JoinDeadlineCountdownProps = {
-  league: IFantasyLeague,
-  onViewLeague: (league: IFantasyLeague) => void
-}
-
-
-function JoinDeadlineCountdown({ league, onViewLeague }: JoinDeadlineCountdownProps) {
-
-  const deadline = calculateJoinDeadline(league);
-
-  if (!deadline) return;
-
-  const diff = epochDiff(deadline);
-
-  const { days, hours, seconds, minutes } = useCountdown(diff);
-  const { userTeam, isLoading, rankedUserTeam } = useUserFantasyTeam(league);
-  const { navigateToMyTeam: navigateToMyTeam } = useRouter();
-
-  const handleCallToAction = () => {
-    if (userTeam) {
-      navigateToMyTeam(userTeam, rankedUserTeam);
-    } else {
-      onViewLeague(league);
-    }
-  }
-
-  const timeBlocks = [
-    { value: days, label: "Days" },
-    { value: hours, label: "Hours" },
-    { value: minutes, label: "Minutes" },
-    { value: seconds, label: "Seconds" }
-  ];
-
-  return (
-    <div className="flex flex-col text-slate-700 dark:text-white bg-white border border-slate-200 dark:border-slate-800 dark:bg-slate-800/50 rounded-xl p-6 gap-4 sm:gap-6">
-      <div className="space-y-2 sm:space-y-4">
-        <h1 className="text-xl sm:text-2xl font-bold tracking-tight">
-          {league.title}
-        </h1>
-        <p className=" dark:text-primary-100  text-sm sm:text-base md:text-lg">
-          Don't miss out on the action. {league.title} starts soon, make sure your team is in!
-        </p>
-      </div>
-
-      <div className="grid grid-cols-4 sm:flex sm:flex-row gap-2 sm:gap-4 items-center justify-start">
-        {timeBlocks.map((block) => (
-          <div
-            key={block.label}
-            className="p-2 sm:p-3 md:min-w-[80px] items-center justify-center flex flex-col rounded-xl bg-slate-50/50 dark:bg-white/10 border border-slate-200 dark:border-white/10"
-          >
-            <p className="font-bold text-lg sm:text-xl md:text-2xl">
-              {block.value.toString().padStart(2, "0")}
-            </p>
-            <p className="text-[10px] sm:text-xs dark:text-primary-100">
-              {block.label}
-            </p>
-          </div>
-        ))}
-      </div>
-
-      <div className="flex items-center gap-4">
-        <button
-          onClick={handleCallToAction}
-          className={twMerge(
-            "w-full sm:w-auto bg-primary-500 dark:bg-primary-600 dark:hover:bg-primary-700 text-primary-50 px-4 sm:px-6 md:px-8 py-2 sm:py-3 rounded-lg font-semibold hover:bg-opacity-90 transition-all flex items-center justify-center sm:justify-start gap-2 shadow-lg",
-            isLoading && "animate-pulse h-10 opacity-30"
-          )}
-        >
-          {!isLoading && (
-            <>
-              <p>{userTeam ? "View Your Team" : "Pick Your Team"}</p>
-              <ChevronRight size={20} />
-            </>
-          )}
-        </button>
-      </div>
-    </div>
-  )
-}
