+++ src/utils/leaguesUtils.tsx
@@ -7,6 +7,11 @@ import { dateComparator } from "./dateUtils";
 export function activeLeaguesFilter(leagues: IFantasyLeague[]) {
   return leagues
     .filter((l) => {
+
+      if (!l.is_open && l.has_ended) {
+        return false;
+      }
+
       if (!l.join_deadline) {
         return false;
       }
@@ -15,7 +20,7 @@ export function activeLeaguesFilter(leagues: IFantasyLeague[]) {
       const deadline = new Date(l.join_deadline);
       const daysDiff = differenceInDays(deadline, today);
 
-      return daysDiff <= 5;
+      return daysDiff <= 7;
     })
     .sort((a, b) => {
       const aDeadline = new Date(a.join_deadline ?? 0);
@@ -151,4 +156,33 @@ export async function latestLeagueFetcher() {
 
     if (sortedLeagues.length === 0) return undefined;
     return sortedLeagues[0];
+}
+
+/** Upcoming Leagues Filter, returns leagues that are more than 7 days away */
+export function upcomingLeaguesFilter(leagues: IFantasyLeague[]) {
+  return leagues
+    .filter((l) => {
+      if (!l.join_deadline) {
+        return false;
+      }
+
+      const today = new Date();
+      const deadline = new Date(l.join_deadline);
+      const daysDiff = differenceInDays(deadline, today);
+
+      return daysDiff > 7;
+    })
+    .sort((a, b) => {
+      const aDeadline = new Date(a.join_deadline ?? 0);
+      const bDealine = new Date(b.join_deadline ?? 0);
+
+      return aDeadline.valueOf() - bDealine.valueOf();
+    });
+}
+
+/** Returns a list of leagues that have ended */
+export function pastLeaguesFilter(leagues: IFantasyLeague[]) {
+  return leagues.filter((l) => {
+    return l.has_ended === true;
+  })
 }
